+++ apps/nextjs/src/components/AppComponents/Chat/chat-lhs-header.tsx
@@ -2,6 +2,7 @@ import React from "react";
 
 import { useRouter } from "next/navigation";
 
+import { useDemoUser } from "@/components/ContextProviders/Demo";
 import { useChatStore } from "@/stores/AilaStoresProvider";
 
 import AiIcon from "../../AiIcon";
@@ -10,20 +11,20 @@ import ChatButton from "./ui/chat-button";
 type ChatLhsHeaderProps = {
   setShowLessonMobile: (value: boolean) => void;
   showLessonMobile: boolean;
-  isDemoUser: boolean;
   showStreamingStatus: boolean;
 };
 
 const ChatLhsHeader = ({
   setShowLessonMobile,
   showLessonMobile,
-  isDemoUser,
   showStreamingStatus,
 }: Readonly<ChatLhsHeaderProps>) => {
   const router = useRouter();
   const ailaStreamingStatus = useChatStore(
     (state) => state.ailaStreamingStatus,
   );
+  const isDemoUser = useDemoUser().isDemoUser;
+
   return (
     <>
       <div className="mt-6 hidden items-center justify-end gap-5 sm:flex">
