+++ apps/nextjs/src/components/AppComponents/Chat/chat-list/working-on-it-message.tsx
@@ -1,7 +1,5 @@
-import {
-  MessageTextWrapper,
-  MessageWrapper,
-} from "@/components/AppComponents/Chat/chat-message";
+import "@/components/AppComponents/Chat/chat-message";
+import { Message } from "@/components/AppComponents/Chat/chat-message/layout";
 import { useChatStore } from "@/stores/AilaStoresProvider";
 
 import { MemoizedReactMarkdownWithStyles } from "../markdown";
@@ -15,12 +13,12 @@ export const WorkingOnItMessage = () => {
   const text = isFinishingUp ? "Finishing up…" : "Working on it…";
 
   return (
-    <MessageWrapper roleType="aila">
-      <MessageTextWrapper>
+    <Message.Container roleType="aila">
+      <Message.Content>
         <div className="w-full animate-pulse">
           <MemoizedReactMarkdownWithStyles markdown={text} />
         </div>
-      </MessageTextWrapper>
-    </MessageWrapper>
+      </Message.Content>
+    </Message.Container>
   );
 };
